+++ Questionable/Controller/GameUi/YesNoChoiceHandler.cs
@@ -5,6 +5,7 @@ using System.Text.RegularExpressions;
 using Dalamud.Game.Addon.Lifecycle;
 using Dalamud.Game.Addon.Lifecycle.AddonArgTypes;
 using Dalamud.Plugin.Services;
+using ECommons.UIHelpers.AddonMasterImplementations;
 using FFXIVClientStructs.FFXIV.Client.UI;
 using FFXIVClientStructs.FFXIV.Component.GUI;
 using Lumina.Excel.Sheets;
@@ -34,6 +35,8 @@ internal sealed class YesNoChoiceHandler : IDisposable
     private readonly IClientState _clientState;
     private readonly ITargetManager _targetManager;
     private readonly ShopController _shopController;
+    private readonly GrandCompanyExchangeController _grandCompanyExchangeController;
+    private readonly ChocoboNamingController _chocoboNamingController;
     private readonly Configuration _configuration;
     private readonly DialogueReferenceResolver _dialogueReferenceResolver;
     private readonly TravelDestinationResolver _travelDestinationResolver;
@@ -54,6 +57,8 @@ internal sealed class YesNoChoiceHandler : IDisposable
         IClientState clientState,
         ITargetManager targetManager,
         ShopController shopController,
+        GrandCompanyExchangeController grandCompanyExchangeController,
+        ChocoboNamingController chocoboNamingController,
         Configuration configuration,
         IDataManager dataManager,
         IPluginLog pluginLog,
@@ -70,6 +75,8 @@ internal sealed class YesNoChoiceHandler : IDisposable
         _clientState = clientState;
         _targetManager = targetManager;
         _shopController = shopController;
+        _grandCompanyExchangeController = grandCompanyExchangeController;
+        _chocoboNamingController = chocoboNamingController;
         _configuration = configuration;
         _dialogueReferenceResolver = dialogueReferenceResolver;
         _travelDestinationResolver = travelDestinationResolver;
@@ -138,6 +145,22 @@ internal sealed class YesNoChoiceHandler : IDisposable
         {
             addonSelectYesno->AtkUnitBase.FireCallbackInt(0);
             _shopController.IsAwaitingYesNo = false;
+            return;
+        }
+
+        if (_grandCompanyExchangeController.IsAwaitingYesNo)
+        {
+            _logger.LogInformation("Confirming Grand Company exchange purchase");
+            new AddonMaster.SelectYesno(addonSelectYesno).Yes();
+            _grandCompanyExchangeController.IsAwaitingYesNo = false;
+            return;
+        }
+
+        if (_chocoboNamingController.IsAwaitingYesNo)
+        {
+            _logger.LogInformation("Confirming chocobo name");
+            new AddonMaster.SelectYesno(addonSelectYesno).Yes();
+            _chocoboNamingController.IsAwaitingYesNo = false;
             return;
         }