+++ Questionable/QuestionablePlugin.cs
@@ -281,6 +281,8 @@ public sealed class QuestionablePlugin : IDalamudPlugin
         serviceCollection.AddSingleton<GatheringController>();
         serviceCollection.AddSingleton<ContextMenuController>();
         serviceCollection.AddSingleton<ShopController>();
+        serviceCollection.AddSingleton<GrandCompanyExchangeController>();
+        serviceCollection.AddSingleton<ChocoboNamingController>();
         serviceCollection.AddSingleton<InterruptHandler>();
 
         serviceCollection.AddSingleton<HighlightObject>();
@@ -376,6 +378,8 @@ public sealed class QuestionablePlugin : IDalamudPlugin
         serviceProvider.GetRequiredService<YesNoChoiceHandler>();
         serviceProvider.GetRequiredService<DialogueChoiceHandler>();
         serviceProvider.GetRequiredService<ShopController>();
+        serviceProvider.GetRequiredService<GrandCompanyExchangeController>();
+        serviceProvider.GetRequiredService<ChocoboNamingController>();
         serviceProvider.GetRequiredService<QuestionableIpc>();
         serviceProvider.GetRequiredService<DalamudInitializer>();
         serviceProvider.GetRequiredService<TextAdvanceIpc>();