+++ Questionable/Windows/JournalComponents/QuestJournalComponent.cs
@@ -33,6 +33,7 @@ internal sealed class QuestJournalComponent
     MovementController movementController,
     AetheryteFunctions aetheryteFunctions,
     AetheryteData aetheryteData,
+    QuestController questController,
     IGameGui gameGui)
 {
     private readonly Dictionary<JournalData.Category, JournalCounts> _categoryCounts = [];
@@ -51,6 +52,7 @@ internal sealed class QuestJournalComponent
     private readonly MovementController _movementController = movementController;
     private readonly AetheryteFunctions _aetheryteFunctions = aetheryteFunctions;
     private readonly AetheryteData _aetheryteData = aetheryteData;
+    private readonly QuestController _questController = questController;
 
     private List<FilteredSection> _filteredSections = [];
 
@@ -244,6 +246,15 @@ internal sealed class QuestJournalComponent
         }
 
         _questJournalUtils.ShowContextMenu(questInfo, quest, nameof(QuestJournalComponent));
+
+        if (quest != null && _questController.PriorityManager.Contains(quest))
+        {
+            ImGui.SameLine();
+            using (_pluginInterface.UiBuilder.IconFontFixedWidthHandle.Push())
+                ImGui.TextColored(ImGuiColors.DalamudYellow, FontAwesomeIcon.ExclamationCircle.ToIconString());
+            if (ImGui.IsItemHovered())
+                ImGui.SetTooltip("This quest is in Priority Quests.");
+        }
 
         ImGui.TableNextColumn();
         float spacing;